requirements.txt
wheel
nbresult
colorama
ipdb
ipykernel
yapf
#matplotlib
pygeohash
pytest
#seaborn
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
google-cloud-bigquery
google-cloud-storage
db-dtypes                  
pyarrow

gputil
psutil
humanize
transformers
torch
glob2
#logging
#pickle5
random2
regex
pytest-shutil

google-cloud
streamlit
setuptools


fastapi         # API framework
pytz            # time zone management
uvicorn         # web server
# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# prevent bq 3+ db-dtypes package error when running training from bq
